refactor: video folder structure
