+++ code/system.ui/react.video/index.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="src/test.ui/entry.tsx"></script>
+    <script type="module" src="src/ui.test/entry.tsx"></script>
   </body>
 </html>
