+++ code/system.ui/react.video/src/index.mts
@@ -1,2 +1,2 @@
 export { Pkg } from './index.pkg.mjs';
-export * from './ui/Vimeo';
+export * from './ui.components/Vimeo';
