+++ code/system.ui/react.video/src/types.mts
@@ -1,1 +1,1 @@
-export * from './ui/Vimeo/types';
+export * from './ui.components/Vimeo/types';
