+++ code/system.ui/react.video/src/ui.components/Vimeo/components/ThumbnailOverlay.tsx
